Move UPnPResultCode to its own file
